backend/accounts/serializers.py
# accounts/serializers.py

from rest_framework import serializers
from .models import CustomUser
from django.contrib.auth.password_validation import validate_password

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ["id", "username", "email", "mobile_no", "is_admin", "is_mobile_verified"]


class UserRegisterSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True, required=True, validators=[validate_password])
    password2 = serializers.CharField(write_only=True, required=True)

    class Meta:
        model = CustomUser
        fields = ["username", "email", "mobile_no", "password", "password2"]

    def validate(self, attrs):
        if attrs["password"] != attrs["password2"]:
            raise serializers.ValidationError({"password": "Password fields didn’t match."})
        return attrs

    def create(self, validated_data):
        # Remove password2 before creating the user
        validated_data.pop("password2")
        user = CustomUser.objects.create_user(**validated_data)
        user.is_active = False  # User will activate after mobile verification
        user.is_mobile_verified = False
        user.save()
        return user
